backend/requirements.txt
ag-ui-protocol==0.1.18
aiofile==3.9.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
argcomplete==3.6.3
attrs==26.1.0
Authlib==1.7.0
beartype==0.22.9
boto3==1.43.1
botocore==1.43.1
cachetools==7.0.6
caio==0.9.25
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cohere==5.21.1
cryptography==47.0.0
cyclopts==4.11.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.18.0
docutils==0.22.4
email-validator==2.3.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.136.1
fastavro==1.12.2
fastmcp==3.2.4
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.4.0
genai-prices==0.0.57
google-auth==2.50.0
google-genai==1.74.0
googleapis-common-protos==1.74.0
griffelib==2.0.2
groq==1.2.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.13.0
idna==3.13
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
Jinja2==3.1.6
jiter==0.14.0
jmespath==1.1.0
joblib==1.5.3
joserfc==1.6.4
jsonpath-python==1.1.5
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.6
jsonschema-specifications==2025.9.1
keyring==25.7.0
logfire==4.32.1
logfire-api==4.32.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.27.0
mdurl==0.1.2
mistralai==2.4.4
more-itertools==11.0.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
nexus-rpc==1.4.0
numpy==2.4.4
openai==2.33.0
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
packaging==25.0
pathable==0.5.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
py-key-value-aio==0.4.4
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-ai==1.89.0
pydantic-ai-slim==1.89.0
pydantic-evals==1.89.0
pydantic-graph==1.89.0
pydantic-handlebars==0.1.0
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
PyMuPDF==1.27.2.3
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==15.0.0
rich-rst==1.3.2
rpds-py==0.30.0
s3transfer==0.17.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sse-starlette==3.4.1
starlette==1.0.0
sympy==1.14.0
temporalio==1.27.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.7.0
typer==0.25.1
types-protobuf==6.32.1.20260221
types-requests==2.33.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
uncalled-for==0.3.1
urllib3==2.6.3
uvicorn==0.46.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
wrapt==1.17.3
xai-sdk==1.12.1
yarl==1.23.0
zipp==3.23.1
